refactor: isolate goose state machine integrations
