+++ crates/goose/src/agents/state_machine/mod.rs
@@ -5,6 +5,7 @@
 //! `StateMachine::run`. Goose's concrete operations remain internal because their
 //! configuration is part of `Agent::reply`, not the state-machine protocol.
 
+mod effects;
 mod machine;
 mod operation;
 mod ops_bang_shell;
@@ -25,15 +26,19 @@ mod ops_tool_approval;
 mod ops_tool_pair_compaction;
 mod ops_toolcalling;
 mod ops_unknown_tool;
+mod session;
+pub(crate) use session::run as run_goose;
 mod usage;
 
 #[cfg(test)]
 mod tests;
 
+pub use effects::GooseEffect;
 pub use machine::{EffectHandler, EffectUsage, MachineSession, SessionLoader, StateMachine, Step};
 pub use operation::{
-    applied, not_applicable, yielded, yielded_with, ConversationEffect, Emitter, GooseEffect,
-    Inference, InferenceInput, MachineEffect, Operation, OperationResult, SlashCommand, StepResult,
+    applied, assistant_turn_count, ends_turn, last_effective_role, messages_since_kickoff,
+    not_applicable, trailing_error, yielded, yielded_with, ConversationEffect, Emitter, Inference,
+    InferenceInput, MachineEffect, Operation, OperationResult, SlashCommand, StepResult,
 };
 
 pub(super) use ops_bang_shell::{bang_shell_command, BangShellOperation};
