+++ crates/goose/src/agents/state_machine/ops_project.rs
@@ -3,7 +3,7 @@
 use anyhow::Result;
 use async_trait::async_trait;
 
-use crate::agents::state_machine::operation::GooseEffect;
+use crate::agents::state_machine::effects::GooseEffect;
 use crate::agents::state_machine::Operation;
 use crate::conversation::Conversation;
 use crate::session::Session;
