+++ crates/goose/src/agents/state_machine/ops_slash_command.rs
@@ -5,8 +5,8 @@ use std::sync::Arc;
 use anyhow::Result;
 use async_trait::async_trait;
 
-use crate::agents::state_machine::operation::GooseEffect;
-use crate::agents::state_machine::operation::{
+use crate::agents::state_machine::effects::GooseEffect;
+use crate::agents::state_machine::{
     messages_since_kickoff, not_applicable, Emitter, Operation, OperationResult, SlashCommand,
 };
 use crate::conversation::Conversation;
