+++ crates/goose/src/agents/state_machine/ops_tool_pair_compaction.rs
@@ -6,11 +6,11 @@ use anyhow::Result;
 use async_trait::async_trait;
 use tracing_futures::Instrument;
 
-use crate::agents::state_machine::operation::{
+use crate::agents::state_machine::ops_llm::chat_span;
+use crate::agents::state_machine::{
     applied, messages_since_kickoff, not_applicable, ConversationEffect, Emitter, GooseEffect,
     Operation, OperationResult,
 };
-use crate::agents::state_machine::ops_llm::chat_span;
 use crate::context_mgmt::{summarize_tool_call, tool_ids_to_summarize};
 use crate::conversation::message::MessageContent;
 use crate::conversation::Conversation;
